chore(deps): Upgrade jol-core to latest
## Description:
To update jol-core some other elements need to be upgraded, particularly
slice. Code has to be updated to work with 64 bit class layout sizes.

+++ presto-main-base/src/main/java/com/facebook/presto/operator/aggregation/differentialentropy/UnweightedDoubleReservoirSample.java
@@ -30,7 +30,7 @@ public class UnweightedDoubleReservoirSample
         implements Cloneable
 {
     public static final int MAX_SAMPLES_LIMIT = 1_000_000;
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(UnweightedDoubleReservoirSample.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(UnweightedDoubleReservoirSample.class).instanceSize();
 
     private int seenCount;
     private double[] samples;
